create BaseTestClass
