+++ twitter_bot/tests/BuyersTest.php
@@ -3,29 +3,10 @@
 namespace TwitterBot\tests;
 
 use PHPUnit\DbUnit\DataSet\YamlDataSet;
-use PHPUnit\DbUnit\TestCase;
-use PHPUnit\DbUnit\TestCaseTrait;
 use TwitterBot\models\Buyers;
-use PDO;
 use Exception;
 
-class BuyersTest extends TestCase {
-
-    use TestCaseTrait;
-
-    private static $pdo = null;
-    private $connection = null;
-
-    public function getConnection() {
-        if ($this->connection == null) {
-            if (self::$pdo == null) {
-                self::$pdo = new PDO($GLOBALS['DB_DSN'], $GLOBALS['DB_USER'], $GLOBALS['DB_PASSWORD']);
-            }
-            $this->connection = $this->createDefaultDBConnection(self::$pdo, $GLOBALS['DB_NAME']);
-        }
-
-        return $this->connection;
-    }
+class BuyersTest extends BaseTestClass {
 
     public function getDataSet() {
         return new YamlDataSet(dirname(__FILE__) .'/files/buyers.yml');
