+++ twitter_bot/tests/JumpsTest.php
@@ -2,29 +2,10 @@
 
 namespace TwitterBot\tests;
 
-use PDO;
-use PHPUnit\Framework\TestCase;
-use PHPUnit\DbUnit\TestCaseTrait;
 use PHPUnit\DbUnit\DataSet\YamlDataSet;
 use TwitterBot\models\Jumps;
 
-class JumpsTest extends TestCase {
-
-    use TestCaseTrait;
-
-    private static $pdo = null;
-    private $connection = null;
-
-    public function getConnection() {
-        if ($this->connection == null) {
-            if (self::$pdo == null) {
-                self::$pdo = new PDO($GLOBALS['DB_DSN'], $GLOBALS['DB_USER'], $GLOBALS['DB_PASSWORD']);
-            }
-            $this->connection = $this->createDefaultDBConnection(self::$pdo, $GLOBALS['DB_NAME']);
-        }
-
-        return $this->connection;
-    }
+class JumpsTest extends BaseTestClass {
 
     public function getDataSet() {
         return new YamlDataSet(dirname(__FILE__).'/files/tables.yml');
